refactor(api): split handler and decorator files
Refs: #25
