+++ api/component.go
@@ -1,14 +1,16 @@
 package api
 
-import "github.com/bwmarrin/discordgo"
+import (
+	"github.com/bwmarrin/discordgo"
+)
 
 // LifecycleHooks allow to specify functions that should be called
 // when components get loaded and unloaded.
 //
 // The defined functions allow some way of initializing a owner.
 type LifecycleHooks struct {
-    LoadComponent   func(discord *discordgo.Session) error
-    UnloadComponent func(discord *discordgo.Session) error
+	LoadComponent   func(discord *discordgo.Session) error
+	UnloadComponent func(discord *discordgo.Session) error
 }
 
 // State holds the state of a owner.
@@ -16,8 +18,8 @@ type LifecycleHooks struct {
 //   - is the owner enabled?
 //   - is the owner currently loaded?
 type State struct {
-    Loaded  bool
-    Enabled bool
+	Loaded  bool
+	Enabled bool
 }
 
 // Component is the base structure that must be aliased
@@ -25,47 +27,47 @@ type State struct {
 //
 // It holds basic metadata about the owner
 type Component struct {
-    // Metadata
-    Name         string
-    Description  string
-    DmPermission bool
+	// Metadata
+	Name         string
+	Description  string
+	DmPermission bool
 
-    // State
-    State State
+	// State
+	State State
 
-    // Lifecycle hooks
-    Lifecycle LifecycleHooks
+	// Lifecycle hooks
+	Lifecycle LifecycleHooks
 
-    // Utilities
-    // These are private and only managed by the API system.
-    // Their initialization happens through call to the methods
-    // used to get them (Example: logger -> Component.Logger()).
-    logger         Logger
-    handlerManager ComponentHandlerManager
-    discord        *discordgo.Session
+	// Utilities
+	// These are private and only managed by the API system.
+	// Their initialization happens through call to the methods
+	// used to get them (Example: logger -> Component.Logger()).
+	logger         Logger
+	handlerManager ComponentHandlerManager
+	discord        *discordgo.Session
 }
 
 // RegistrableComponent is the interface that allows a owner to be
 // initialized and registered.
 type RegistrableComponent interface {
-    RegisterComponent(discord *discordgo.Session) error
-    UnregisterComponent(discord *discordgo.Session) error
+	RegisterComponent(discord *discordgo.Session) error
+	UnregisterComponent(discord *discordgo.Session) error
 }
 
 // RegisterComponent is used by the owner registration system that
 // automatically calls the RegisterComponent method for all Component instances in
 // the components.Components array.
 func (c *Component) RegisterComponent(discord *discordgo.Session) error {
-    c.discord = discord
+	c.discord = discord
 
-    err := c.Lifecycle.LoadComponent(discord)
+	err := c.Lifecycle.LoadComponent(discord)
 
-    if err != nil {
-        return err
-    }
-    c.State.Loaded = true
+	if err != nil {
+		return err
+	}
+	c.State.Loaded = true
 
-    return nil
+	return nil
 }
 
 // UnregisterComponent is used by the owner registration system that
@@ -74,12 +76,12 @@ func (c *Component) RegisterComponent(discord *discordgo.Session) error {
 //
 // It is used to give components the ability to gracefully shutdown.
 func (c *Component) UnregisterComponent(discord *discordgo.Session) error {
-    err := c.Lifecycle.UnloadComponent(discord)
+	err := c.Lifecycle.UnloadComponent(discord)
 
-    if err != nil {
-        return err
-    }
-    c.State.Loaded = false
+	if err != nil {
+		return err
+	}
+	c.State.Loaded = false
 
-    return nil
+	return nil
 }
